+++ packages/terra-date-time-picker/src/DateTimePicker.jsx
@@ -81,6 +81,10 @@ const propTypes = {
    * An ISO 8601 string representation of the initial value to show in the date and time inputs.
    */
   value: PropTypes.string,
+  /**
+   * Type of time input to initialize. Must be '24-hour' or '12-hour'
+   */
+  timeVariant: PropTypes.oneOf([DateTimeUtils.FORMAT_12_HOUR, DateTimeUtils.FORMAT_24_HOUR]),
 };
 
 const defaultProps = {
@@ -97,6 +101,7 @@ const defaultProps = {
   requestFocus: undefined,
   timeInputAttributes: undefined,
   value: undefined,
+  timeVariant: DateTimeUtils.FORMAT_24_HOUR,
 };
 
 class DateTimePicker extends React.Component {
@@ -404,6 +409,7 @@ class DateTimePicker extends React.Component {
       releaseFocus,
       timeInputAttributes,
       value,
+      timeVariant,
       ...customProps
     } = this.props;
 
@@ -451,6 +457,7 @@ class DateTimePicker extends React.Component {
             name="input"
             value={this.timeValue}
             disabled={disabled}
+            variant={timeVariant}
           />
 
           {this.state.isAmbiguousTime ? this.renderTimeClarification() : null }